converso_telegram_bot/pyproject.toml
# FILEPATH: /home/gianfranco/Desktop/uni/masters-thesis/converso_telegram_bot/pyproject.toml

[tool.poetry]
name = "converso_telegram_bot"
version = "0.1.0"
description = "Artificial Intelligence assistant - Telegram Bot"
authors = ["Gianfranco Demarco <gianfrademarco@gmail.com>"]

[tool.poetry.group.lint.dependencies]
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10, <4.0"

"aio_pika" = "9.3.1"
"pika" = "1.3.2"
"python-telegram-bot" = "20.7"
"requests" = "2.31.0"
"redis" = "5.0.1"
"openai" = "1.3.6"